+++ RobotServer/controllers/http_robot/http_robot.py
@@ -14,6 +14,9 @@ import threading
 import time
 import traceback
 
+from device_map import get_device_id
+from motor_requests import apply_motor_requests
+
 
 # flask log level
 log = logging.getLogger("werkzeug")
@@ -37,13 +40,7 @@ def create_app(device_map, motor_requests, robot):
     @app.route("/motors", methods=["PUT"])
     def put_motors():
         request_data = request.json
-        for request_motor_values in request_data["motors"]:
-            request_motor_id = request_motor_values.get("id")
-            motor = device_map["Motors"].get(request_motor_id)
-            if motor:
-                motor_requests[request_motor_id] = request_motor_values
-            else:
-                raise Exception(f"No motor named {request_motor_id} found")
+        apply_motor_requests(request_data["motors"], device_map, motor_requests)
 
         # return sensor data
 
@@ -135,6 +132,8 @@ def create_app(device_map, motor_requests, robot):
                     )
                 ),
                 # "WorldPose": {"Position": position, "Yaw": yaw, "Time": time},
+                # Some static fake data to not break the api
+                "WorldPose": {"Position": 0, "Yaw": 0, "Time": time},
             }
         )
 
@@ -196,62 +195,6 @@ def create_app(device_map, motor_requests, robot):
         return "OK"
 
     return app
-
-
-def build_device_map(robot, timestep) -> dict:
-    device_map = defaultdict(dict)
-
-    device_count = robot.getNumberOfDevices()
-    print(f"Found {device_count} devices on robot")
-    for i in range(device_count):
-        device = robot.getDeviceByIndex(i)
-        device_type = device.getNodeType()
-        device_id = get_device_id(device)
-
-        if device_type == Node.ROTATIONAL_MOTOR or device_type == Node.LINEAR_MOTOR:
-            # Initialize the motor with an infinite target position so that we can directly control velocity
-            device.setPosition(float("inf"))
-            device.setVelocity(0)
-            device_map["Motors"][device_id] = device
-        elif device_type == Node.DISTANCE_SENSOR:
-            # Initialize the distance sensor with an update frequency
-            device.enable(timestep)
-            device_map["DistanceSensors"][device_id] = device
-        elif (
-            device_type == Node.TOUCH_SENSOR and device.getType() == TouchSensor.BUMPER
-        ):
-            device.enable(timestep)
-            device_map["BumperTouchSensor"][device_id] = device
-        elif device_type == Node.POSITION_SENSOR:
-            device.enable(timestep)
-            device_map["PositionSensors"][device_id] = device
-            # Handle PositionSensorLimitSwitch
-            device_node = robot.getFromDevice(device)
-            if (
-                device_node
-                and device_node.isProto()
-                and device_node.getTypeName() == "PositionSensorLimitSwitch"
-            ):
-                device_map["PositionSensorLimitSwitch"][device_id] = device
-        elif device_type == Node.INERTIAL_UNIT:
-            device.enable(timestep)
-            device_map["IMUs"][device_id] = device
-        elif device_type == Node.CAMERA:
-            device.enable(timestep)
-            device_map["Cameras"][device_id] = device
-        elif device_type == Node.GYRO:
-            device.enable(timestep)
-            device_map["Gyros"][device_id] = device
-        else:
-            print(
-                f"Found unknown device of type {device_type} with ID {device_id} not mapped"
-            )
-
-    return device_map
-
-
-def get_device_id(device: str):
-    return device.getName().split("#")[0].strip()
 
 
 def get_public_ip():
