+++ RobotServer/controllers/http_robot/tests/conftest.py
@@ -14,6 +14,7 @@ def test_app():
     motor_requests = {}
     device_map = collections.defaultdict(dict)
     robot = MagicMock()
+    robot.getTime.return_value = 0
     robot.getSelf.getOrientation.return_value = [0, 0]
     app = create_app(device_map=device_map, motor_requests=motor_requests, robot=robot)
     with app.test_client() as test_client:
